backend/core/serializers/user_serializer.py
from core.models.user import User
from rest_framework import serializers



class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'first_name', 'last_name', 'email', 'institution', 'nick_name')
