backend/requirements.txt
django
pylint
pylint-django
pyjwt
djangorestframework
django-extensions

pillow

numpy
bs4
pdfkit
